refactor: edit view test
